requirements.txt
django
pytest
python-decouple
mysqlclient
pyyaml
ua-parser
user-agents
django-user-agents
django-messages
coverage
djangorestframework